+++ sdk/schemaregistry/azure-schemaregistry-avroencoder/azure/schemaregistry/encoder/avroencoder/_schema_registry_avro_encoder.py
@@ -130,6 +130,7 @@ class AvroEncoder(object):
         *,
         schema: str,
         message_type: Optional[Callable] = None,
+        request_kwargs: Dict[str, Any] = None,
         **kwargs: Any,
     ) -> Union[MessageType, MessageMetadataDict]:
         """
@@ -154,6 +155,8 @@ class AvroEncoder(object):
          `(data: bytes, content_type: str, **kwargs) -> MessageType`, where `data` and `content_type`
          are positional parameters.
         :paramtype message_type: Callable or None
+        :keyword request_kwargs: The keyword arguments to be passed to the client.
+        :paramtype request_kwargs: Dict[str, Any]
         :rtype: MessageType or MessageMetadataDict
         :raises ~azure.schemaregistry.encoder.avroencoder.exceptions.SchemaParseError:
             Indicates an issue with parsing schema.
@@ -170,7 +173,8 @@ class AvroEncoder(object):
                 f"Cannot parse schema: {raw_input_schema}", error=e
             ).raise_with_traceback()
 
-        schema_id = self._get_schema_id(schema_fullname, raw_input_schema)
+        request_kwargs = request_kwargs or {}
+        schema_id = self._get_schema_id(schema_fullname, raw_input_schema, **request_kwargs)
         content_type = f"{AVRO_MIME_TYPE}+{schema_id}"
 
         try:
@@ -225,6 +229,7 @@ class AvroEncoder(object):
         message: Union[MessageType, MessageMetadataDict],
         *,
         readers_schema: Optional[str] = None,
+        request_kwargs: Dict[str, Any] = None,
         **kwargs,   # pylint: disable=unused-argument
     ) -> Dict[str, Any]:
         """
@@ -241,6 +246,8 @@ class AvroEncoder(object):
         :type message: MessageType or MessageMetadataDict
         :keyword readers_schema: An optional reader's schema as defined by the Apache Avro specification.
         :paramtype readers_schema: str or None
+        :keyword request_kwargs: The keyword arguments to be passed to the client.
+        :paramtype request_kwargs: Dict[str, Any]
         :rtype: Dict[str, Any]
         :raises ~azure.schemaregistry.encoder.avroencoder.exceptions.SchemaParseError:
             Indicates an issue with parsing schema.
@@ -267,7 +274,8 @@ class AvroEncoder(object):
         data, content_type = self._convert_preamble_format(data, content_type)
 
         schema_id = content_type.split("+")[1]
-        schema_definition = self._get_schema(schema_id)
+        request_kwargs = request_kwargs or {}
+        schema_definition = self._get_schema(schema_id, **request_kwargs)
         try:
             dict_value = self._avro_encoder.decode(
                 data, schema_definition, readers_schema=readers_schema