+++ sdk/schemaregistry/azure-schemaregistry-avroencoder/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 1.0.0b2 (Unreleased)
+
+### Features Added
+
+- `request_kwargs` has been added to `encode` and `decode` on `AvroEncoder` as an optional parameter to be passed into client requests.
+- The size of the current schema/schema ID caches will be logged at an info level when a new entry has been added.
+
 ## 1.0.0b1 (2022-02-09)
 
 This version and all future versions will require Python 3.6+. Python 2.7 is no longer supported.